Code Cleanup
This commit cleans up the code to improve its readability and maintainability:

- **Refactor `main.py`**: The `main.py` file has been refactored to improve its structure and readability. The business logic has been moved to separate modules.
- **Refactor `App.js`**: The `App.js` file has been refactored to improve its structure and readability. The component has been broken down into smaller, more manageable components.
- **Add comments and docstrings**: Comments and docstrings have been added to the code to explain the purpose of each function and component.
